# specs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Project Chimera API",
    "description": "Autonomous Influencer Network Platform",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/trends/": {
      "get": {
        "tags": [
          "trends"
        ],
        "summary": "List Trends",
        "description": "Fetch real trends via NewsData.io and cache.",
        "operationId": "list_trends_api_v1_trends__get",
        "parameters": [
          {
            "name": "niche",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Topic e.g. fashion, crypto",
              "title": "Niche"
            },
            "description": "Topic e.g. fashion, crypto"
          },
          {
            "name": "time_window",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "4h, 24h, 7d, 30d",
              "default": "24h",
              "title": "Time Window"
            },
            "description": "4h, 24h, 7d, 30d"
          },
          {
            "name": "location",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Country code e.g. US",
              "title": "Location"
            },
            "description": "Country code e.g. US"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trends": {
      "get": {
        "tags": [
          "trends"
        ],
        "summary": "List Trends",
        "description": "Fetch real trends via NewsData.io and cache.",
        "operationId": "list_trends_api_v1_trends_get",
        "parameters": [
          {
            "name": "niche",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Topic e.g. fashion, crypto",
              "title": "Niche"
            },
            "description": "Topic e.g. fashion, crypto"
          },
          {
            "name": "time_window",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "4h, 24h, 7d, 30d",
              "default": "24h",
              "title": "Time Window"
            },
            "description": "4h, 24h, 7d, 30d"
          },
          {
            "name": "location",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Country code e.g. US",
              "title": "Location"
            },
            "description": "Country code e.g. US"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/agents/": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "List Agents",
        "description": "List agents (stub).",
        "operationId": "list_agents_api_v1_agents__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentResponse"
                  },
                  "type": "array",
                  "title": "Response List Agents Api V1 Agents  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Create Agent",
        "description": "Create agent (stub).",
        "operationId": "create_agent_api_v1_agents__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "List Agents",
        "description": "List agents (stub).",
        "operationId": "list_agents_api_v1_agents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentResponse"
                  },
                  "type": "array",
                  "title": "Response List Agents Api V1 Agents Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Create Agent",
        "description": "Create agent (stub).",
        "operationId": "create_agent_api_v1_agents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{agent_id}": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Get Agent",
        "description": "Get agent by id (stub).",
        "operationId": "get_agent_api_v1_agents__agent_id__get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/content/": {
      "get": {
        "tags": [
          "content"
        ],
        "summary": "List Content",
        "description": "List content (stub).",
        "operationId": "list_content_api_v1_content__get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContentResponse"
                  },
                  "title": "Response List Content Api V1 Content  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "content"
        ],
        "summary": "Create Content",
        "description": "Create content (stub).",
        "operationId": "create_content_api_v1_content__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/content": {
      "get": {
        "tags": [
          "content"
        ],
        "summary": "List Content",
        "description": "List content (stub).",
        "operationId": "list_content_api_v1_content_get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContentResponse"
                  },
                  "title": "Response List Content Api V1 Content Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "content"
        ],
        "summary": "Create Content",
        "description": "Create content (stub).",
        "operationId": "create_content_api_v1_content_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commerce/": {
      "get": {
        "tags": [
          "commerce"
        ],
        "summary": "List Transactions",
        "description": "List transactions (stub).",
        "operationId": "list_transactions_api_v1_commerce__get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionResponse"
                  },
                  "title": "Response List Transactions Api V1 Commerce  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "commerce"
        ],
        "summary": "Create Transaction",
        "description": "Create transaction (stub).",
        "operationId": "create_transaction_api_v1_commerce__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commerce": {
      "get": {
        "tags": [
          "commerce"
        ],
        "summary": "List Transactions",
        "description": "List transactions (stub).",
        "operationId": "list_transactions_api_v1_commerce_get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionResponse"
                  },
                  "title": "Response List Transactions Api V1 Commerce Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "commerce"
        ],
        "summary": "Create Transaction",
        "description": "Create transaction (stub).",
        "operationId": "create_transaction_api_v1_commerce_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "persona_id": {
            "type": "string",
            "title": "Persona Id"
          },
          "niche": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Niche"
          }
        },
        "type": "object",
        "required": [
          "name",
          "persona_id"
        ],
        "title": "AgentCreate"
      },
      "AgentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "persona_id": {
            "type": "string",
            "title": "Persona Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "persona_id"
        ],
        "title": "AgentResponse"
      },
      "ContentCreate": {
        "properties": {
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type",
            "default": "text"
          },
          "platform": {
            "type": "string",
            "title": "Platform",
            "default": "twitter"
          },
          "text_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Content"
          }
        },
        "type": "object",
        "required": [
          "agent_id"
        ],
        "title": "ContentCreate"
      },
      "ContentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "draft"
          }
        },
        "type": "object",
        "required": [
          "id",
          "agent_id",
          "content_type",
          "platform"
        ],
        "title": "ContentResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "TransactionCreate": {
        "properties": {
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "transaction_type": {
            "type": "string",
            "title": "Transaction Type",
            "default": "transfer"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "to_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Address"
          },
          "purpose": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purpose"
          }
        },
        "type": "object",
        "required": [
          "agent_id",
          "amount"
        ],
        "title": "TransactionCreate"
      },
      "TransactionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "transaction_type": {
            "type": "string",
            "title": "Transaction Type"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "pending"
          }
        },
        "type": "object",
        "required": [
          "id",
          "agent_id",
          "transaction_type",
          "amount"
        ],
        "title": "TransactionResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}